pyproject.toml
[project]
name = "kami-excel-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "openai>=2.24.0",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.2",
]
